requirements.txt
streamlit
qrcode
firebase-admin
google-cloud-storage
pillow
